steven: fix collision bugs
The direction vector in checkCollisions uses the camera as the previous
position. Since the camera wasn't updated until after all the collision
checks were done the direction vector would still include the X
component for the Z check and the X and Z components for the Y check.
This caused the client to move more that it should have out of the block
and glitch into other blocks.