steven: wait until the first packet until starting the writer
There seems to be a race condition in the vanilla server which means
for a short while after LoginSuccess its still in the LOGIN state.
This means any packets we send instantly (like arm swings which can
happen early) will be looked up in the LOGIN state's packets instead
of PLAY causing the server to disconnect us. Waiting until the first
packet from the server ensures its in the right state.